style : toast popup, directory update
