+++ src/main/java/com/ohgiraffers/springeagles/global/auth/entity/UserEntity.java
@@ -1,0 +1,37 @@
+package com.ohgiraffers.springeagles.global.auth.entity;
+
+import com.ohgiraffers.springeagles.global.auth.model.UserRole;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.Set;
+
+@Entity
+@Table(name = "global_users")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class UserEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_id", nullable = false)
+    private Integer userId;
+
+    @Column(name = "user_name", nullable = false)
+    private String userName;
+
+    @Column(name = "user_password", nullable = false)
+    private String userPassword;
+
+    @Column(name = "user_email", nullable = false)
+    private String userEmail;
+
+    @ElementCollection(fetch = FetchType.EAGER)
+    @CollectionTable(name = "global_user_roles", joinColumns = @JoinColumn(name = "user_id"))
+    @Column(name = "user_roles", nullable = false)
+    @Enumerated(EnumType.STRING)
+    private Set<UserRole> roles;
+}
