+++ src/main/java/com/ohgiraffers/springeagles/global/auth/model/UserDTO.java
@@ -1,0 +1,16 @@
+package com.ohgiraffers.springeagles.global.auth.model;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class UserDTO {
+    private Integer userID;
+    private String userName;
+    private String userPassword;
+    private String userEmail;
+    private String role;
+}
