+++ src/main/java/com/ohgiraffers/springeagles/global/auth/util/PasswordEncoderConfig.java
@@ -1,0 +1,17 @@
+package com.ohgiraffers.springeagles.global.auth.util;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class PasswordEncoderConfig {
+
+    // 순환 참조를 막기위해 따로 분리하여 구성
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
