--- src/main/java/com/ohgiraffers/springeagles/global/index/Post.java
@@ -1,61 +1,0 @@
-package com.ohgiraffers.springeagles.global.index;
-
-public class Post {
-    private String title;
-    private String description;
-    private String imageUrl;
-    private String link;
-
-    public Post(String title, String description, String imageUrl, String link) {
-        this.title = title;
-        this.description = description;
-        this.imageUrl = imageUrl;
-        this.link = link;
-    }
-
-    public Post() {
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public void setImageUrl(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-
-    public String getLink() {
-        return link;
-    }
-
-    public void setLink(String link) {
-        this.link = link;
-    }
-
-    @Override
-    public String toString() {
-        return "Post{" +
-                "title='" + title + '\'' +
-                ", description='" + description + '\'' +
-                ", imageUrl='" + imageUrl + '\'' +
-                ", link='" + link + '\'' +
-                '}';
-    }
-
-}
