+++ src/main/java/com/ohgiraffers/springeagles/global/index/controller/SessionController.java
@@ -1,0 +1,14 @@
+package com.ohgiraffers.springeagles.global.index.controller;
+
+import jakarta.servlet.http.HttpSession;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class SessionController {
+
+    @PostMapping("/clear-login-success")
+    public void clearLoginSuccess(HttpSession session) {
+        session.removeAttribute("loginSuccess");
+    }
+}
