--- src/main/java/com/ohgiraffers/springeagles/global/user/config/SecurityConfig.java
@@ -1,67 +1,0 @@
-package com.ohgiraffers.springeagles.global.user.config;
-
-import com.ohgiraffers.springeagles.global.error.CustomAccessDeniedHandler;
-import com.ohgiraffers.springeagles.global.user.service.CustomUserDetailsService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.method.configuration.GlobalMethodSecurityConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-public class SecurityConfig extends GlobalMethodSecurityConfiguration {
-
-    private final CustomUserDetailsService userDetailsService;
-    private final PasswordEncoder passwordEncoder;
-    private final CustomAccessDeniedHandler accessDeniedHandler;
-
-    @Autowired
-    public SecurityConfig(CustomUserDetailsService userDetailsService, PasswordEncoder passwordEncoder,
-                          CustomAccessDeniedHandler accessDeniedHandler) {
-        this.userDetailsService = userDetailsService;
-        this.passwordEncoder = passwordEncoder;
-        this.accessDeniedHandler = accessDeniedHandler;
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .csrf(AbstractHttpConfigurer::disable)
-                .authorizeRequests(authorizeRequests ->
-                        authorizeRequests
-                                .requestMatchers("/signup", "/css/**", "/javascript/**", "/images/**").permitAll()
-                                .requestMatchers("/user/**").hasAuthority("ROLE_USER")
-                                .requestMatchers("/admin/**").hasAuthority("ROLE_ADMIN")
-                                .anyRequest().authenticated())
-                .formLogin(formLogin ->
-                        formLogin
-                                .loginPage("/login")
-                                .defaultSuccessUrl("/", true)
-                                .permitAll())
-                .logout(logout ->
-                        logout
-                                .logoutUrl("/logout")
-                                .logoutSuccessUrl("/login?logout")
-                                .permitAll())
-                .exceptionHandling(exceptionHandling ->
-                        exceptionHandling
-                                .accessDeniedHandler(accessDeniedHandler));
-
-        return http.build();
-    }
-
-    @Autowired
-    public void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userDetailsService)
-                .passwordEncoder(passwordEncoder);
-    }
-}
