--- src/main/java/com/ohgiraffers/springeagles/global/user/dto/UserDTO.java
@@ -1,16 +1,0 @@
-package com.ohgiraffers.springeagles.global.user.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class UserDTO {
-    private Integer userID;
-    private String userName;
-    private String userPassword;
-    private String userEmail;
-    private String role;
-}
