--- src/main/java/com/ohgiraffers/springeagles/global/user/dto/UserRole.java
@@ -1,24 +1,0 @@
-package com.ohgiraffers.springeagles.global.user.dto;
-
-public enum UserRole {
-
-    USER("ROLE_USER"),
-    ADMIN("ROLE_ADMIN");
-
-    private String role;
-
-    UserRole(String role) {
-        this.role = role;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    @Override
-    public String toString() {
-        return "UserRole{" +
-                "role='" + role + '\'' +
-                '}';
-    }
-}
