--- src/main/java/com/ohgiraffers/springeagles/global/user/service/CustomUserDetailsService.java
@@ -1,37 +1,0 @@
-package com.ohgiraffers.springeagles.global.user.service;
-
-import com.ohgiraffers.springeagles.global.user.repository.UserEntity;
-import com.ohgiraffers.springeagles.global.user.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-import java.util.Optional;
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Autowired
-    public CustomUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        UserEntity user = userRepository.findByUserName(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        return new User(user.getUserName(), user.getUserPassword(),
-                Collections.singletonList(new SimpleGrantedAuthority(user.getUserRole().getRole())));
-    }
-
-    public Optional<UserEntity> findByUserName(String username) {
-        return userRepository.findByUserName(username);
-    }
-}
