+++ src/main/java/com/ohgiraffers/springeagles/jstBlog/comment/model/STCommentDTO.java
@@ -1,0 +1,20 @@
+package com.ohgiraffers.springeagles.jstBlog.comment.model;
+
+import lombok.*;
+
+import java.time.LocalDate;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class STCommentDTO {
+
+    private Integer commentId;
+    private String commentContent;
+    private LocalDate commentDate;
+    private LocalDate commentModifyTime;
+    private Integer postId;
+    private String userName;
+}
