--- src/main/java/com/ohgiraffers/springeagles/jstBlog/likes/dto/STLikesDTO.java
@@ -1,14 +1,0 @@
-package com.ohgiraffers.springeagles.jstBlog.likes.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class STLikesDTO {
-    private Integer likesId;
-    private Integer userId;
-    private Integer postId;
-}
