+++ src/main/java/com/ohgiraffers/springeagles/jstBlog/likes/service/STLikesService.java
@@ -1,11 +1,11 @@
 
 package com.ohgiraffers.springeagles.jstBlog.likes.service;
 
-import com.ohgiraffers.springeagles.global.user.repository.UserEntity;
-import com.ohgiraffers.springeagles.global.user.repository.UserRepository;
-import com.ohgiraffers.springeagles.jstBlog.likes.repository.STLikesEntity;
+import com.ohgiraffers.springeagles.global.auth.entity.UserEntity;
+import com.ohgiraffers.springeagles.global.auth.repository.UserRepository;
+import com.ohgiraffers.springeagles.jstBlog.likes.entity.STLikesEntity;
 import com.ohgiraffers.springeagles.jstBlog.likes.repository.STLikesRepository;
-import com.ohgiraffers.springeagles.jstBlog.posts.repository.STPostsEntity;
+import com.ohgiraffers.springeagles.jstBlog.posts.entity.STPostsEntity;
 import com.ohgiraffers.springeagles.jstBlog.posts.repository.STPostsRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
