+++ src/main/java/com/ohgiraffers/springeagles/jstBlog/posts/controller/STPostsController.java
@@ -1,12 +1,12 @@
 package com.ohgiraffers.springeagles.jstBlog.posts.controller;
 
-import com.ohgiraffers.springeagles.global.user.repository.UserEntity;
-import com.ohgiraffers.springeagles.global.user.service.CustomUserDetailsService;
-import com.ohgiraffers.springeagles.jstBlog.comment.repository.STCommentEntity;
+import com.ohgiraffers.springeagles.global.auth.entity.UserEntity;
+import com.ohgiraffers.springeagles.global.auth.service.CustomUserDetailsService;
+import com.ohgiraffers.springeagles.jstBlog.comment.entity.STCommentEntity;
 import com.ohgiraffers.springeagles.jstBlog.comment.service.STCommentService;
 import com.ohgiraffers.springeagles.jstBlog.likes.service.STLikesService;
-import com.ohgiraffers.springeagles.jstBlog.posts.dto.STPostsDTO;
-import com.ohgiraffers.springeagles.jstBlog.posts.repository.STPostsEntity;
+import com.ohgiraffers.springeagles.jstBlog.posts.model.STPostsDTO;
+import com.ohgiraffers.springeagles.jstBlog.posts.entity.STPostsEntity;
 import com.ohgiraffers.springeagles.jstBlog.posts.service.STPostsService;
 import com.ohgiraffers.springeagles.jstBlog.userIntro.service.UserIntroService;
 import org.springframework.beans.factory.annotation.Autowired;
