+++ src/main/java/com/ohgiraffers/springeagles/jstBlog/posts/repository/STPostsRepository.java
@@ -1,5 +1,6 @@
 package com.ohgiraffers.springeagles.jstBlog.posts.repository;
 
+import com.ohgiraffers.springeagles.jstBlog.posts.entity.STPostsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
