+++ src/main/java/com/ohgiraffers/springeagles/jstBlog/userIntro/repository/UserIntroRepository.java
@@ -1,5 +1,6 @@
 package com.ohgiraffers.springeagles.jstBlog.userIntro.repository;
 
+import com.ohgiraffers.springeagles.jstBlog.userIntro.entity.UserIntroEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
