+++ src/main/resources/templates/index.html
@@ -9,6 +9,7 @@
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
     <script th:src="@{/javascript/mainPage.js}"></script>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/sweetalert2@11.4.10/dist/sweetalert2.min.css">
 </head>
 <body>
 <header>
@@ -18,7 +19,7 @@
         </div>
         <div class="header-btn">
             <div class="header-authentication" th:text="${username} + ' 님 환영합니다'"></div>
-            <div  class="header-logout" onclick="logout()">로그아웃</div>
+            <div class="header-logout" onclick="logout()">로그아웃</div>
             <form id="searchForm" action="search" method="get">
                 <div class="search-box">
                     <button type="submit" class="btn-search" id="searchButton">
@@ -187,7 +188,8 @@
         </div>
     </div>
 </section>
-<script>
+<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.4.10/dist/sweetalert2.min.js"></script>
+<script th:inline="javascript">
     // Swiper 객체 생성 및 옵션 설정
     const swiper = new Swiper(".swiper-container", {
         grabCursor: true,
@@ -210,10 +212,43 @@
 
     function logout() {
         if (confirm("로그아웃 하시겠습니까?")) {
-            alert("로그아웃 되었습니다.")
             location.href = "/logout";
         } else {
             return;
+        }
+    }
+
+    // 로그인 성공 메시지 처리
+    window.onload = function() {
+        /*[[${session.loginSuccess}]]*/;
+        const loginSuccess = /*[[${session.loginSuccess}]]*/ false;
+        console.log("Login Success:", loginSuccess); // Debugging line
+
+        if (loginSuccess) {
+            Swal.fire({
+                toast: true,
+                position: 'top-end',
+                icon: 'success',
+                title: '로그인에 성공하였습니다!',
+                showConfirmButton: false,
+                timer: 3000,
+                timerProgressBar: true,
+                customClass: {
+                    popup: 'custom-font'
+                },
+                didOpen: (toast) => {
+                    toast.addEventListener('mouseenter', Swal.stopTimer)
+                    toast.addEventListener('mouseleave', Swal.resumeTimer)
+                }
+            }).then(() => {
+                fetch('/clear-login-success', {
+                    method: 'POST'
+                }).then(response => {
+                    console.log('Session attribute cleared');
+                }).catch(error => {
+                    console.error('Error clearing session attribute:', error);
+                });
+            });
         }
     }
 </script>
