+++ src/main/resources/templates/login.html
@@ -7,8 +7,6 @@
     <link rel="stylesheet" th:href="@{/css/public_index.css}" />
     <!-- SweetAlert2 CSS -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/sweetalert2@11.4.10/dist/sweetalert2.min.css">
-    <style>
-    </style>
 </head>
 <body>
 <div class="login-align">
@@ -43,6 +41,7 @@
         /* Error case */
         const errorParam = /*[[${param.error}]]*/ false;
         const logoutParam = /*[[${param.logout}]]*/ false;
+        const successMessage = /*[[${successMessage}]]*/ false;
 
         if (errorParam) {
             Swal.fire({
@@ -69,6 +68,25 @@
                 position: 'top-end',
                 icon: 'success',
                 title: '성공적으로 로그아웃 되었습니다',
+                showConfirmButton: false,
+                timer: 3000,
+                timerProgressBar: true,
+                customClass: {
+                    popup: 'custom-font'
+                },
+                didOpen: (toast) => {
+                    toast.addEventListener('mouseenter', Swal.stopTimer)
+                    toast.addEventListener('mouseleave', Swal.resumeTimer)
+                }
+            });
+        }
+
+        if (successMessage) {
+            Swal.fire({
+                toast: true,
+                position: 'top-end',
+                icon: 'success',
+                title: '회원가입이 완료되었습니다',
                 showConfirmButton: false,
                 timer: 3000,
                 timerProgressBar: true,
